### 2019-08-01 370be9a
remove spaces from ref labels
