+++ paper.bib
@@ -17,7 +17,7 @@
     year={1992},
     pages={461–469}}
 
-@article{hees_fang_langford_assah_mohammad_silva_trenell_white_wareham_brage_et al.:2014,
+@article{hees_fang_langford_assah_mohammad_silva_trenell_white_wareham_brage_etal.:2014,
     title={Autocalibration of accelerometer data for free-living physical activity assessment using local gravity and temperature: an evaluation on four continents},
     volume={117},
     DOI={10.1152/japplphysiol.00421.2014},
@@ -27,7 +27,7 @@
     year={2014},
     pages={738–744}}
 
-@article{hees_sabia_anderson_denton_oliver_catt_abell_kivimäki_trenell_singh-manoux_et al.:2015,
+@article{hees_sabia_anderson_denton_oliver_catt_abell_kivimäki_trenell_singh-manoux_etal.:2015,
     title={A Novel, Open Access Method to Assess Sleep Duration Using a Wrist-Worn Accelerometer},
     volume={10},
     DOI={10.1371/journal.pone.0142533},
@@ -36,7 +36,7 @@
     author={Hees, Vincent T. Van and Sabia, Séverine and Anderson, Kirstie N. and Denton, Sarah J. and Oliver, James and Catt, Michael and Abell, Jessica G. and Kivimäki, Mika and Trenell, Michael I. and Singh-Manoux, Archana and et al.},
     year={2015}}
 
-@article{hees_sabia_jones_wood_anderson_kivimäki_frayling_pack_bucan_trenell_et al.:2018,
+@article{hees_sabia_jones_wood_anderson_kivimäki_frayling_pack_bucan_trenell_etal.:2018,
     title={Estimating sleep parameters using an accelerometer without sleep diary},
     volume={8},
     DOI={10.1038/s41598-018-31266-z},
