### 2019-08-01 #1617 @labarba
@elyiorgos  — Please merge this fix
https://github.com/elyiorgos/sleeppy/pull/1
