### 2019-08-01 #1617 @labarba
Please merge this fix
https://github.com/elyiorgos/sleeppy/pull/2
