### 2019-08-02 a08466e
Reformatted bibtex file using mendeley to resolve references issue
