+++ paper.bib
@@ -1,54 +1,78 @@
-@article{bai_di_xiao_evenson_lacroix_crainiceanu_buchner:2017,
-    title={An Activity Index For Raw Accelerometry Data And Its Application In Older Adults},
-    volume={1}, DOI={10.1093/geroni/igx004.4497},
-    number={suppl_1},
-    journal={Innovation in Aging},
-    author={Bai, J. and Di, C. and Xiao, L. and Evenson, K.r. and Lacroix, A. and Crainiceanu, C. and Buchner, D.m.},
-    year={2017},
-    pages={1239–1239}}
-
-@article{cole_kripke_gruen_mullaney_gillin:1992,
-    title={Automatic Sleep/Wake Identification From Wrist Activity},
-    volume={15},
-    DOI={10.1093/sleep/15.5.461},
-    number={5},
-    journal={Sleep},
-    author={Cole, Roger J. and Kripke, Daniel F. and Gruen, William and Mullaney, Daniel J. and Gillin, J. Christian},
-    year={1992},
-    pages={461–469}}
-
-@article{hees_fang_langford_assah_mohammad_silva_trenell_white_wareham_brage_etal.:2014,
-    title={Autocalibration of accelerometer data for free-living physical activity assessment using local gravity and temperature: an evaluation on four continents},
-    volume={117},
-    DOI={10.1152/japplphysiol.00421.2014},
-    number={7},
-    journal={Journal of Applied Physiology},
-    author={Hees, Vincent T. Van and Fang, Zhou and Langford, Joss and Assah, Felix and Mohammad, Anwar and Silva, Inacio C. M. Da and Trenell, Michael I. and White, Tom and Wareham, Nicholas J. and Brage, Søren and et al.},
-    year={2014},
-    pages={738–744}}
-
-@article{hees_sabia_anderson_denton_oliver_catt_abell_kivimäki_trenell_singh-manoux_etal.:2015,
-    title={A Novel, Open Access Method to Assess Sleep Duration Using a Wrist-Worn Accelerometer},
-    volume={10},
-    DOI={10.1371/journal.pone.0142533},
-    number={11},
-    journal={Plos One},
-    author={Hees, Vincent T. Van and Sabia, Séverine and Anderson, Kirstie N. and Denton, Sarah J. and Oliver, James and Catt, Michael and Abell, Jessica G. and Kivimäki, Mika and Trenell, Michael I. and Singh-Manoux, Archana and et al.},
-    year={2015}}
-
-@article{hees_sabia_jones_wood_anderson_kivimäki_frayling_pack_bucan_trenell_etal.:2018,
-    title={Estimating sleep parameters using an accelerometer without sleep diary},
-    volume={8},
-    DOI={10.1038/s41598-018-31266-z},
-    number={1},
-    journal={Scientific Reports},
-    author={Hees, Vincent Theodoor Van and Sabia, S. and Jones, S. E. and Wood, A. R. and Anderson, K. N. and Kivimäki, M. and Frayling, T. M. and Pack, A. I. and Bucan, M. and Trenell, M. I. and et al.},
-    year={2018}}
-
-@misc{hees_fang_zhao_heywood_mirkes_sabia_migueles:2019,
-    title = {GGIR: Raw Accelerometer Data Analysis},
-    DOI={10.5281/zenodo.1051064},
-    url={https://crantastic.org/packages/GGIR},
-    year={2019},
-    author={Hees, V.T. Van and Fang, Z. and Zhao, J. and Heywood, J. and Mirkes, E. and Sabia, S. and Migueles, J.}
+@article{VanHees2014,
[…]
