### 2019-08-02 8addfa5
Reformatted paper.md to produce proper pdf output
