+++ paper.md
@@ -42,45 +42,60 @@ Measures of sleep quality and quantity can provide valuable insights into the he
 ``SleepPy`` has been developed with the goal of making it easy for researchers to derive measures of sleep from accelerometer data collected during daily life. While it currently provides out-of-the-box support for data files generated by the GeneActiv wrist-worn devices, it can be easily extended to support data files generated by other wrist-worn wearable devices. The package can work with both GeneActiv .bin files, as well as raw .csv outputs generated by GeneActiv's software. Each .bin file needs to be manually processed with the GeneActiv software to generate a .csv file. So while it takes longer for ``SleepPy`` to process .bin files, it is highly advisable to do so for large datasets.
 
 ``SleepPy`` follows seven steps when processing data:
+
 1.	Split data by day: Load raw accelerometer data from input file and split it into 24-hour segments (noon to noon).
-2.	Derive activity index -`[@bai_di_xiao_evenson_lacroix_crainiceanu_buchner:2017]`: Calculate activity index for each 1 minute epoch of the day.
-3.	Perform off-body detection -`[@hees_fang_zhao_heywood_mirkes_sabia_migueles:2019]`: Run off-body detection algorithm and generate a label (on or off body) for each 15 minute epoch of the day.
-4.	Identify major rest period -`[@hees_sabia_jones_wood_anderson_kivimäki_frayling_pack_bucan_trenell_et al.:2018]`: Estimate the major rest period (i.e. sleep window) for each day.
-5.	Perform sleep/wake classification -`[@cole_kripke_gruen_mullaney_gillin:1992]`: Run sleep/wake classification algorithm to generate sleep/wake labels for each 1 minute epoch of the day.
+
+2.	Derive activity index [@Bai2016]: Calculate activity index for each 1 minute epoch of the day.
+
+3.	Perform off-body detection [@VanHees2019]: Run off-body detection algorithm and generate a label (on or off body) for each 15 minute epoch of the day.
+
+4.	Identify major rest period [@VanHees2018]: Estimate the major rest period (i.e. sleep window) for each day.
+
+5.	Perform sleep/wake classification [@Cole1992]: Run sleep/wake classification algorithm to generate sleep/wake labels for each 1 minute epoch of the day.
+
 6.	Calculate sleep measures: Calculate sleep measures based on sleep/wake states only during the major rest period for each day.
+
 7.	Generate reports and visualizations: Create a set of tables and charts for analysis and presentation of processed outputs and sleep measures for each day.
 
 # Outputs
 The user has access to all intermediate data, a table of sleep measures, and two kinds of visual reports.
 
-![Demo Report](https://raw.githubusercontent.com/elyiorgos/sleeppy/master/sleeppy/demo/report_images/Visual_Results_Day_1.png)
-
-Figure 1: Visual report showing various raw and processed data streams
+![Visual report showing various raw and processed data streams](https://raw.githubusercontent.com/elyiorgos/sleeppy/master/sleeppy/demo/report_images/Visual_Results_Day_1.png)
 
 # Data Streams Report
 Raw and processed data streams generated at each stage of the processing pipeline are visualized in this report for each day (24 hour period, noon to noon). As shown in Figure 1, the report includes a source file name, number of the day in the sequence of days provided, start date of the data being shown, and a table of all calculated sleep measures for that day. Below the table are plots of raw and processed data streams.
 
 The plots are laid out as follows:
+
 1.	XYZ: The first chart is a plot of the raw tri-axial accelerometer signals (X, Y, and Z).
+
 2.	Activity index: The second chart is a plot of minute-by-minute activity index values, which reflect the intensity of activity for each minute.
+
[…]
