### 2019-08-02 cf53c17
One more fix for references, make sure the references heading is at the end of the file
