### 2019-08-03 #1617 @labarba
👋  @lheagy — Could you handle this JOSS submission as editor? (You might get @fperez to review it: I know he was capturing data from his sleep at one point!)

**SleepPy: A python package for sleep analysis from accelerometer data**
https://github.com/elyiorgos/sleeppy
[ Article proof :page_facing_up:](https://github.com/openjournals/joss-papers/blob/joss.01617/joss.01617/10.21105.joss.01617.pdf)
