### 2019-08-06 #1617 @lheagy
Thanks @elyiorgos. Before assigning reviewers, there are a few things that I wasn't able to find in the repository that are a part of the [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html). If you would be willing to address the following before we start the review process, this should help streamline the review:

- [ ] [Community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) for other communities interested in contributing to the project
- [ ] [Tests](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines): is there a test-suite that covers the core functionality of the software? (Ideally automated).

Also, for documentation, you might consider hosting the API documentation on [ReadTheDocs](https://readthedocs.org/) or similar. This is not required but a suggestion.

For reference, here is an accepted JOSS paper that is particularly well-done: https://joss.theoj.org/papers/10.21105/joss.00957
