+++ CONTRIBUTING.md
@@ -1,0 +1,23 @@
+# Contributing
+We encourage any and all contributions to the project. 
+If you want to contribute, please [fork the master branch](https://github.com/elyiorgos/sleeppy/fork) and open a pull request with your edits. 
+We will reply to pull requests and are more than happy to discuss any problems or address any questions you may have about the project or process.
+
+## Documentation
+
+Improving documentation is a great way to make first contributions. 
+If you found the wording of any of the documentation unclear, or the flow of the documentation difficult to follow, 
+please feel free to suggest improvements to make the package more user friendly.
+
+## Code
+
+If you wish to add new features or improve existing code, we are happy to review and approve your suggested changes or
+updates. Please just ensure that your addition is well explained, well justified, and clearly commented so that its
+purpose and function are evident to the reviewer. 
+
+## Issues
+
+If you find any issues, discover any errors, or can not make the package run feel free to [file an issue](https://github.com/elyiorgos/sleeppy/issues/new). 
+When filing your issue please be as detailed and explicit as possible, include the libraries you are using and their versions, the operating system you are using, and any other information you feel is relevant. 
+Please also include as much information as necessary to reproduce the problem you encountered.
+In the event that you find an issue and are able to resolve it on your own, please submit a pull request with the updated code to help improve the package for others.
