### 2019-08-07 #1617 @elyiorgos
Hi @lheagy, I just added community guidelines today after looking at the example you sent, and a few other accepted joss examples, I appreciate the guidance :)

As a way to test the function of the package, there is a demo file and demo run script provided in the sleeppy folder. Essentially what this does is take an acceptable input data file, and run the full package on it, and produce all intermediate data as well as several visualizations to confirm things are functioning properly. As a way to check that the outputs are appropriate outputs, the expected visual results are also stored in the demo folder. Running the test/demo is outlined in the readme and the documentation on pip, but I'm also happy to answer any questions.

I wanted to ask, is it customary to bring reviewers from an external network with a JOSS submission? We have several contacts at various institutions/companies in our network that work in wearable technologies and/or would be potential users of the package, so we could suggest them as reviewers if that's acceptable or desirable. Let me know and I'm happy to suggest extra reviewers!

Thanks very much for all the help!

Best,
Yiorgos
