### 2019-08-09 #1617 @lheagy
Thanks @elyiorgos!

One note on testing, we do expect that there is at least some testing that objectively checks the code (e.g. a sample input and `assert` statements or similar that numerically verify the results). Tests that require a visual check aren't enough to meet [those criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#tests)

As for reviewers, yes - it is acceptable to bring in external reviewers. We do expect that reviewers have some experience developing software so that they can provide input from a developer's on the software package. If you do have people in mind who you would be interested in getting feedback from or making a connection with, this is a great opportunity, so feel free to share their names!

All the best,
Lindsey
