### 2019-08-09 cc2da0a
Added more objective test case to demo, updated versioning to reflect changes, updated documentation in readme to include a warning about automatic requirement installation (specific versions) and added suggestion to use virtual environments(with a helpful link)
