+++ README.md
@@ -2,12 +2,24 @@
 A Python(2.7) package for sleep analysis from accelerometer data
 
 ## Overview
-Measures of sleep quality and quantity can provide valuable insights into the health and well-being of an individual. Traditionally, sleep assessments are performed in the clinic/hospital setting using polysomnography tests. Recent advances in wearable sensor technology have enabled objective assessment of sleep at home. Actigraphy has been widely used for this purpose and several algorithms have been published in the literature over the years. However, implementation of these algorithms is not widely available, which creates a barrier for wider adoption of wearable devices in clinical research.
+Measures of sleep quality and quantity can provide valuable insights into the health and well-being of an individual. 
+Traditionally, sleep assessments are performed in the clinic/hospital setting using polysomnography tests. 
+Recent advances in wearable sensor technology have enabled objective assessment of sleep at home. 
+Actigraphy has been widely used for this purpose and several algorithms have been published in the literature 
+over the years. However, implementation of these algorithms is not widely available, which creates a barrier for wider 
+adoption of wearable devices in clinical research.
 
-``SleepPy`` is an open source python package incorporating several published algorithms in a modular framework and providing a suite of measures for the assessment of sleep quantity and quality. The package can process multi-day streams of raw accelerometer data (X, Y & Z) from wrist-worn wearable devices to produce sleep reports and visualizations for each recording day (24-hour period). The reports are formatted to facilitate statistical analysis of sleep measures. Visualization acts as a quick debugging tool, provides insights into sleep patterns of individual subjects and can be used for presentation of data to diverse audiences.
+``SleepPy`` is an open source python package incorporating several published algorithms in a modular framework and 
+providing a suite of measures for the assessment of sleep quantity and quality. The package can process multi-day 
+streams of raw accelerometer data (X, Y & Z) from wrist-worn wearable devices to produce sleep reports and 
+visualizations for each recording day (24-hour period). The reports are formatted to facilitate statistical 
+analysis of sleep measures. Visualization acts as a quick debugging tool, provides insights into sleep patterns of 
+individual subjects and can be used for presentation of data to diverse audiences.
 
 ## Requirements
-``SleepPy`` can be installed through pip, which will handle requirement installation for you:
+``SleepPy`` can be installed through pip, which will handle requirement installation for you (please note that SleepPy 
+will use specific versions of all required packages, to avoid possible issues with your existing environments 
+we suggest using a [virtual environment](https://packaging.python.org/guides/installing-using-pip-and-virtual-environments/)):
 
 ```sh
 pip install sleeppy
@@ -41,18 +53,26 @@ pip install -r requirements.txt
 
 2.	Derive activity index: Calculate activity index for each 1 minute epoch of the day.
 
-3.	Perform off-body detection: Run off-body detection algorithm and generate a label (on or off body) for each 15 minute epoch of the day.
+3.	Perform off-body detection: Run off-body detection algorithm and generate a label (on or off body) for each 15 
+minute epoch of the day.
 
 4.	Identify major rest period: Estimate the major rest period (i.e. sleep window) for each day.
 
-5.	Perform sleep/wake classification: Run sleep/wake classification algorithm to generate sleep/wake labels for each 1 minute epoch of the day.
+5.	Perform sleep/wake classification: Run sleep/wake classification algorithm to generate sleep/wake labels for each 
+1 minute epoch of the day.
 
-6.	Calculate sleep measures: Calculate sleep measures based on sleep/wake states only during the major rest period for each day.
+6.	Calculate sleep measures: Calculate sleep measures based on sleep/wake states only during the major rest period for 
+each day.
 
[…]
