### 2019-08-09 #1617 @elyiorgos
Thanks @lheagy!

The existing demo produces a visual report that includes calculated
endpoints(their numerical values), which could be compared with the
expected endpoints manually, but based on your response and for ease of use
I added an automated check for that process today. That should serve as a
more robust validation of the output.

To help identify external reviewers I'm going to include my colleague
@shyamalpatel on this thread, we should have a list of potential reviewers
ready soon (hopefully today). We'll share that as soon as we get in touch
with them. Thanks again!

Best,
Yiorgos

On Fri, Aug 9, 2019 at 12:10 AM Lindsey Heagy <notifications@github.com>
wrote:

> Thanks @elyiorgos <https://github.com/elyiorgos>!
>
> One note on testing, we do expect that there is at least some testing that
> objectively checks the code (e.g. a sample input and assert statements or
> similar that numerically verify the results). Tests that require a visual
> check aren't enough to meet those criteria
> <https://joss.readthedocs.io/en/latest/review_criteria.html#tests>
>
> As for reviewers, yes - it is acceptable to bring in external reviewers.
> We do expect that reviewers have some experience developing software so
> that they can provide input from a developer's on the software package. If
> you do have people in mind who you would be interested in getting feedback
> from or making a connection with, this is a great opportunity, so feel free
> to share their names!
>
> All the best,
> Lindsey
>
