### 2019-08-09 #1617 @shyamalpatel
We would like to nominate Prof. Bjoern Eskofier (@eskofier) as a reviewer. Dr. Eskofier is a professor in the Department of Computer Science at University of Erlangen (https://www.fau.eu/) and leads a research group focused on wearable computing and machine learning.

Best,
Shyamal
