### 2019-08-09 080aca7
Used python black to reformat all relevant files in sleeppy
