+++ sleeppy/sleep.py
@@ -2,8 +2,9 @@ import pandas as pd
 import os
 import numpy as np
 import matplotlib
+
 try:
-    matplotlib.use('TkAgg')
+    matplotlib.use("TkAgg")
 except Exception as e:
     print "Error: could not use TkAgg as backend"
     pass
@@ -13,9 +14,10 @@ import seaborn as sns
 import struct
 from bitstring import BitArray
 from scipy import signal
+
 sns.set()
 pd.options.mode.chained_assignment = None
-__all__ = ['SleepPy', 'ColeKripke', 'band_pass_filter', 'activity_index', 'bin2df']
+__all__ = ["SleepPy", "ColeKripke", "band_pass_filter", "activity_index", "bin2df"]
 
 
 class SleepPy:
@@ -60,8 +62,17 @@ class SleepPy:
 
 
     """
-    def __init__(self, input_file, results_directory, sampling_frequency, start_buffer='0s', stop_buffer='0s',
-                 start_time='', stop_time=''):
+
+    def __init__(
+        self,
+        input_file,
+        results_directory,
+        sampling_frequency,
+        start_buffer="0s",
+        stop_buffer="0s",
+        start_time="",
+        stop_time="",
+    ):
         """
         Class initialization.
 
@@ -75,11 +86,16 @@ class SleepPy:
         """
         self.src = input_file  # save input location
         self.dst = results_directory  # save output location
-        self.src_name = input_file.split('/')[-1][0:-4]  # save naming convention
-        self.sub_dst = results_directory + '/' + self.src_name  # create output directory
+        self.src_name = input_file.split("/")[-1][0:-4]  # save naming convention
+        self.sub_dst = (
+            results_directory + "/" + self.src_name
+        )  # create output directory
         self.fs = sampling_frequency  # save sampling frequency
         self.window_size = 60  # define window size in seconds
-        self.band_pass_cutoff = (0.25, 12.0)  # define the cutoffs for the band pass filter
+        self.band_pass_cutoff = (
+            0.25,
+            12.0,
+        )  # define the cutoffs for the band pass filter
         self.major_rest_periods = []  # initialize a list to save the major rest periods
         self.start_buffer = start_buffer
         self.stop_buffer = stop_buffer
@@ -97,9 +113,9 @@ class SleepPy:
         except OSError:
             pass
         # split the data into 24 hour periods
-        if '.bin' in self.src:
+        if ".bin" in self.src:
             self.split_days_geneactiv_bin()
-        elif '.csv' in self.src:
+        elif ".csv" in self.src:
             self.split_days_geneactiv_csv()
 
         # extract the activity index feature
@@ -126,54 +142,72 @@ class SleepPy:
 
         """
         try:
-            os.mkdir(self.sub_dst + '/raw_days')  # set up output directory
+            os.mkdir(self.sub_dst + "/raw_days")  # set up output directory
         except OSError:
             pass
         # load data and fix time_stamps
-        data = pd.read_csv(self.src,
-                           index_col=0,
-                           skiprows=100,
-                           header=None,
-                           names=['Time', 'X', 'Y', 'Z', 'LUX', 'Button', 'T'],
-                           usecols=['Time', 'X', 'Y', 'Z', 'T'],
-                           dtype={'Time': object,
-                                  'X': np.float64,
-                                  'Y': np.float64,
-                                  'Z': np.float64,
-                                  'LUX': np.int64,
-                                  'Button': bool,
-                                  'T': np.float64},
-                           low_memory=False)
+        data = pd.read_csv(
+            self.src,
+            index_col=0,
[…]
