### 2019-08-09 1995219
Reformatted all print statements for eventual python 3 integration
