### 2019-08-12 #1617 @lheagy
:wave: Hi @eskofier, @jminnier would you be willing to review this submission [JOSS](Joss.theoj.org/about)? JOSS is a 'developer friendly' journal. That is, we focus on the software as a scientific contribution. If the submitting authors have followed best practices (have documentation, tests, continuous integration, and a license) then their review should be straight-forward. There are more details on reviewing with JOSS here: http://joss.theoj.org/about#reviewer_guidelines

I would greatly appreciate your input! If you are unable to take this on at the moment, any recommendations for alternate reviewers would be much appreciated.
