### 2019-08-12 #1617 @eskofier
Dear Lindsay,

I can take this.

Best regards,
Bjoern Eskofier
--
Prof. Bjoern Eskofier, Ph.D.

FAU Sandbox: find out more about our newest FAU student startup & innovation program at www.sandbox.fau.de

Heisenberg professorship of the DFG
Digital Sports and Health Group
Machine Learning and Data Analytics Lab
Department of Computer Science
Faculty of Engineering
Friedrich-Alexander University Erlangen-Nuernberg (FAU)
Carl-Thiersch-Strasse 2b, 91052 Erlangen, Germany

Adjunct Member, Faculty of Medicine
Speaker, Central Institute of Healthcare Engineering
Friedrich-Alexander University Erlangen-Nuernberg (FAU)
Henkestraße 91, 91052 Erlangen

I-net: http://www.mad.cs.fau.de
Phone: +49 9131 85-27297
Email: bjoern.eskofier@fau.de

> Am 12.08.2019 um 06:01 schrieb Lindsey Heagy <notifications@github.com>:
>
> 👋 Hi @eskofier, @jminnier would you be willing to review this submission JOSS? JOSS is a 'developer friendly' journal. That is, we focus on the software as a scientific contribution. If the submitting authors have followed best practices (have documentation, tests, continuous integration, and a license) then their review should be straight-forward. There are more details on reviewing with JOSS here: http://joss.theoj.org/about#reviewer_guidelines
>
> I would greatly appreciate your input! If you are unable to take this on at the moment, any recommendations for alternate reviewers would be much appreciated.
>
