### 2019-08-12 #1617 @shyamalpatel
@lheagy [Dr. Ikaro Silva](https://www.linkedin.com/in/ikaro-silva-5642239/) (@ikarosilva) would be able to review this if you are looking for additional reviewers. Dr. Silva is a senior research scientist at Philips Research North America and has a strong background in biomedical applications of signal processing and machine learning.

Best,
Shyamal
