### 2019-08-15 #1617 @jminnier
Sorry @lheagy I don't have time for the next few weeks to review this one, though this package looks really neat! Hope you are able to find another reviewer.
