### 2019-08-19 #1663 @whedon
[REVIEW]: SleepPy: A python package for sleep analysis from accelerometer data
**Submitting author:** @elyiorgos (<a href="http://orcid.org/0000-0003-3371-8935">Yiorgos Christakis</a>)
**Repository:** <a href="https://github.com/elyiorgos/sleeppy" target ="_blank">https://github.com/elyiorgos/sleeppy</a>
**Version:** 0.2.21
