### 2019-08-20 #1663 @lheagy
Thank you @eskofier and @ikarosilva for being willing to review! Above there is a checklist to help guide your review. Please feel free to open issues on the [target repository](https://github.com/elyiorgos/sleeppy) with comments or suggestions for improvement in the submission and link them here by mentioning `openjournals/joss-reviews#1663` in the issue text.

If you are able to complete your review in the next two weeks, that would be much appreciated. Please let me know if you have any questions along the way. Thanks again!
