### 2019-09-11 #1663 @eskofier
Hi Kevin,

What exactly are we supposed to do? I have not seen any review instructions or the likes?

Best regards,
Bjoern

-sent from mobile office-

> Am 11.09.2019 um 06:23 schrieb Kevin Mattheus Moerman <notifications@github.com>:
>
> @eskofier, @ikarosilva thanks again for your help with this review. Can you please give us an indication as to when you can work on/finalize this review? Thanks.
>
