### 2019-09-17 #1663 @ikarosilva
Hi @lheagy,

Posting here per per our discussion on conflict of interest waiver. I have worked with Dr. Patel for several months and  a year on a team at a start up in 2015. I hold him at a very high regards and as a trusted colleague. We also have a few publications in press. While I don't believe our relationship should affect my ability to review this work, I would like to disclose this information in spirit of full transparency.
