### 2019-09-29 #1663 @nhejazi
@elyiorgos, the `sleeppy` package looks interesting, thanks for submitting the paper to JOSS! After going through the reviewer checklist, there are a few outstanding items that I'd like to bring your attention:

* The package requires a statement about contribution guidelines. A simple approach for doing this includes adding a section "Contributing" to the `README` that links to a more complete document (e.g., `CONTRIBUTING.md`), giving those interested in adding to the package information on how best to do so.
* From an examination of the package source, it appears that there are no unit tests included in the package. There are several frameworks for writing such tests, including [`unittest`](https://docs.python.org/2/library/unittest.html), [`nose2`](https://github.com/nose-devs/nose2), and [`pytest`](https://pytest.readthedocs.io/en/latest/). Such tests are critical for ensuring software integrity as development continues. If I missed a note about tests or they are located elsewhere, feel free to point this out also.
