### 2019-09-29 #1663 @elyiorgos
Hello Nima!

Thank you for your comments!

In the current sleeppy github repo there should be an existing
CONTRIBUTING.md file that outlines the ways in which any interested party
can contribute if they would like to. If that file needs more fleshing out
I'm happy to update it. I'm also happy to add a link to the README.md file
to make it easier to find if that is a suitable and helpful addition.

With regards to a testing suite, there is a demo module with a demo.py file
that runs the full suite of analyses on demo data, generates reports to be
compared to the expected output(from a visual point of view), and also
performs assertion tests on the expected clinical endpoints calculated by
the package. My understanding in talking with the editor pre-submission was
that those would suffice, however, I fully understand the importance of
more robust testing, and am more than happy to create a suite of more
conventional unit tests testing the many specific pieces of functionality
leveraged by the package.

When you get a chance let me know your thoughts on both accounts, and I'll
get right on it! Thank you again!

Wishing you all the best,
Yiorgos Christakis

On Sun, Sep 29, 2019 at 12:20 AM Nima Hejazi <notifications@github.com>
wrote:

> @elyiorgos <https://github.com/elyiorgos>, the sleeppy package looks
> interesting, thanks for submitting the paper to JOSS! After going through
> the reviewer checklist, there are a few outstanding items that I'd like to
> bring your attention:
>
>    - The package requires a statement about contribution guidelines. A
>    simple approach for doing this includes adding a section "Contributing" to
>    the README that links to a more complete document (e.g.,
>    CONTRIBUTING.md), giving those interested in adding to the package
>    information on how best to do so.
>    - From an examination of the package source, it appears that there are
>    no unit tests included in the package. There are several frameworks for
>    writing such tests, including unittest
>    <https://docs.python.org/2/library/unittest.html>, nose2
>    <https://github.com/nose-devs/nose2>, and pytest
>    <https://pytest.readthedocs.io/en/latest/>. Such tests are critical
>    for ensuring software integrity as development continues. If I missed a
>    note about tests or they are located elsewhere, feel free to point this out
>    also.
>
