### 2019-10-02 06d415b
Added link to contribution doc
