### 2019-10-02 #1663 @elyiorgos
Hi Nima,

Thanks for taking a look at what I mentioned. I went ahead and updated the
the README to include a link to the contributions document. I completely
agree with your testing point, robust unit testing and integration tests
for the project are at the top of our to do list in terms of development.
Thanks again for your help and responsiveness! We're very excited!

Best,
Yiorgos

On Mon, Sep 30, 2019 at 10:24 PM Nima Hejazi <notifications@github.com>
wrote:

> Hi Yiorgos ---
>
> Thanks for your quick reply and for pointing out the items that I missed.
> The included CONTRIBUTING.md file looks good to me --- please do link to
> it from the README.md though as this will make it much easier for
> interested users to find the document. Also, you're right that formal unit
> tests are technically not necessary to pass the JOSS review, that is, the
> checks currently included in the demo.py file appear sufficient. I will
> check off the testing item as it appears that the included checks do
> satisfy the requirement; however, I'd still strongly encourage adding unit
> tests in the style of whichever framework you prefer in the future as this
> will go a long way towards ensuring reproducibility as changes to the core
> software are made (possibly by contributors). Once such tests are set up,
> you may also consider looking into leveraging existing continuous
> integration services to ensure that your tests are run upon each commit /
> PR (you could also set this up with the existing demo.py and perhaps
> check that key numerical results are consistent between commits / PRs).
> Happy to check off the remaining item about contributions as well after the
> README.md is updated. Almost there!
>
> Cheers,
> Nima
>
