### 2019-10-04 #1663 @lheagy
Many thanks @nhejazi for your speedy review!

@elyiorgos: if you are able to include a test suite, that would be preferable for publication. Although having examples run does act as a test, that really is the bare minimum.
