### 2019-10-04 #1663 @lheagy
:wave: Just checking in @eskofier. How are things going? In order to check-off the boxes for the review, you will need to have accepted the JOSS invitation: https://github.com/openjournals/joss-reviews/invitations. Please let me know if you have any questions. Thanks for your work on this!
