### 2019-10-06 #1663 @ikarosilva
Hi Guys, sorry for the delay.
Update my review list. My feedback:

1) I did not see any unit tests (am I missing something)? Would be nice to have a standard unit test folder that uses python unit test frameworks (nose etc).

2) Please avoid the use of wild imports such as:

from sleeppy.sleep import *
Instead, limit the scope of your import to only those functions/classes necessary.

3)
