### 2019-10-07 #1663 @eskofier
Dear all,

I took a look at this with my senior PhD student Stefan Gradl (github: gradlman, also CCed). Feedback:

-> the code looks well documented where needed
-> we could not really test the code because of its programming in python 2.7. I don’t have a 2.7 installation anymore, neither does Stefan or most people in my lab (except 1). Stefan tried to get the code running for half a day by downloading old packages or searching for compatible ones, but without success. As Python 2.7 will not be supported anymore after the end of this year [1, 2], we decided to not try further.

[1] https://pythonclock.org
[2] https://www.anaconda.com/end-of-life-eol-for-python-2-7-is-coming-are-you-ready/

Please let us know how to proceed.

Best regards,
Bjoern Eskofier
--
Prof. Bjoern Eskofier, Ph.D.

FAU Sandbox: find out more about our newest FAU student startup & innovation program at www.sandbox.fau.de

Heisenberg professorship of the DFG
Digital Sports and Health Group
Machine Learning and Data Analytics Lab
Department of Computer Science
Faculty of Engineering
Friedrich-Alexander University Erlangen-Nuernberg (FAU)
Carl-Thiersch-Strasse 2b, 91052 Erlangen, Germany

Adjunct Member, Faculty of Medicine
Speaker, Central Institute of Healthcare Engineering
Friedrich-Alexander University Erlangen-Nuernberg (FAU)
Henkestraße 91, 91052 Erlangen

I-net: http://www.mad.cs.fau.de
Phone: +49 9131 85-27297
Email: bjoern.eskofier@fau.de

> Am 07.10.2019 um 00:34 schrieb Ikaro Silva <notifications@github.com>:
>
> Hi Guys, sorry for the delay.
> Update my review list. My feedback:
>
> 	• I did not see any unit tests (am I missing something)? Would be nice to have a standard unit test folder that uses python unit test frameworks (nose etc).
>
> 	• Please avoid the use of wild imports such as:
>
> from sleeppy.sleep import *
> Instead, limit the scope of your import to only those functions/classes necessary.
>
