### 2019-10-08 #1663 @eskofier
Hi,

pip install did not work on an apple computer

Bjoern$ pip install sleeppy
Collecting sleeppy
  ERROR: Could not find a version that satisfies the requirement sleeppy (from versions: none)
ERROR: No matching distribution found for sleeppy

Manual install did not work

error: Setup script exited with error: library dfftpack has Fortran sources but no Fortran compiler found
(as well as several other errors)

Sorry, but I do not have time to mess with technology. I have a couple of hours of work ahead of me and it’s 10p

Viele Grüße
Bjoern

> Am 08.10.2019 um 21:36 schrieb Yiorgos Christakis <notifications@github.com>:
>
> Hello all,
>
> I'll try to address things in order, though forgive me if I miss something.
>
> @nhejazi Thank you for your quick feedback, and quick responses!
>
> @lheagy With regards to a testing suite, without getting into any unnecessary details, the research project that helped spawn the package is currently being overhauled from a testing point of view, which will eventually require the addition of a full testing suite to the package, though this may take some time as other pipelines take priority.
>
> @ikarosilva Thank you for your comments! I'll be sure to fix the wild import you showed, as well as any other similar instance across the package. As far as testing goes, the demo folder provides dummy data that can be automatically run through the package, producing reports and clinical endpoints for one night's sleep. When the demo is run, there are assertion tests that run through the produced endpoints and confirm that they match the expected output. I realize this is more of a full package integration test, and a full suite of unit tests in addition to that is desirable to verify that the package functions perform as expected, though as I mentioned above that may take some time. As written, the package does meet the minimum testing requirements for publication, though I understand if that causes you(or others) to have reservations in your review.
>
> @eskofier I'm sorry to hear you had difficulty getting the package to install and run. Assuming you have access to an apple computer, python 2.7 should be built-in and the sleeppy package can simply be pip-installed. The pip installation is written in such a way that it installs all required dependencies with the correct versions, which should be all you need to be able to run the demo/test from inside the python interpreter (more detailed run directions are available in the readme). I'm happy to provide any assistance beyond that if necessary, please let me know.
>
> A conversion of the package to python 3 will soon be in the works (for exactly the reasons you mentioned), but will take a back seat to fleshing out testing, so the timeline for that will be somewhat more extended. Our choice of python 2.7 was related to the fact that the package wasn't developed in a vacuum, but had to be integrated with a large existing project code base that had python 2.7 dependencies.
>
> Please let me know if there are any more questions or comments, and thank you all for your help evaluating the package.
>
> Best,
> Yiorgos Christakis
>
