### 2019-10-12 #1663 @eskofier
Hi Yiorgos,

Did not work :/ When I run:

$ pip2 install sleeppy
-bash: pip2: command not found

Also:

$ pip3 install sleeppy
Collecting sleeppy
  ERROR: Could not find a version that satisfies the requirement sleeppy (from versions: none)
ERROR: No matching distribution found for sleeppy

Like before

$ pip install sleeppy
Collecting sleeppy
  ERROR: Could not find a version that satisfies the requirement sleeppy (from versions: none)
ERROR: No matching distribution found for sleeppy

What should we try next?

Viele Grüße
Bjoern

> Am 11.10.2019 um 17:54 schrieb Yiorgos Christakis <notifications@github.com>:
>
> Hi Bjoern,
>
> I understand your frustration, and I'm sorry this is adding to an already demanding schedule. My guess is that since your laptop is probably rather new and you yourself use python 3, it's likely that the default python environment is python 3, so when you run a normal pip command like you showed above, it tries to pull a distribution for python 3 and can't find one. I actually ran into the same issue on my personal laptop, but if you run pip2 install sleeppy it should download and set up without a hitch. On my personal I had to run sudo pip2 install sleeppy due to a permissions error. After that, if you choose to run it from the python interpreter, just make sure that when starting python in the command line you run python2 as opposed to just python. Again that should work, but if not please let me know. If necessary I'm more than happy to set up a call/chat with either you or your colleague to sort things out. Thank you again for agreeing to review the package, and for your feedback.
>
> Best,
> Yiorgos
>
