### 2019-10-14 #1663 @elyiorgos
Hi Bjoern,

Are you able to run python 2 from the command line with the command `python2`? If so you can install pip2 by running the following:

This will get you what you need to download pip
`wget https://bootstrap.pypa.io/get-pip.py`

This will install pip for python 2
 `sudo python2 get-pip.py`

This will install pip for python 3
`sudo python3 get-pip.py`

Now if you call pip, or pip3, you will get pip for python 3, and if you call pip2, you'll get pip for python 2. From there running `pip2 install sleeppy` should work.

If running `python2` doesn't work for you for whatever reason, let me know and we can continue from there.

Best,
Yiorgos
