### 2019-10-14 #1663 @lheagy
@elyiorgos and @eskofier: If getting up and running locally is proving to be challenging, then I would recommend that @elyiorgos set up an example that runs on [binder](http://mybinder.org). You will need to specify an environment.yml file that contains the dependencies, and binder will install those for you and give you a sandbox computational environment to run the code. See for example this demo repository: https://github.com/binder-examples/python-conda_pip.

Creating a binder demonstrates that the software can be installed, and doesn't mess with @elyiorgos's software environment.

Please let me know if you run into any challenges getting up and running with binder.
