### 2019-10-15 #1663 @yiorg
@lheagy Thank you for the suggestion, that seems like the best way forward. I'll get working on it as soon as possible and share it here once I have it up and running!

Best,
Yiorg
