### 2019-10-19 d6b318f
added pytables to requirements
