### 2019-10-21 2df1cf2
Syncing external development updates, fixing example notebook, created binder specific demo version, added files to facilitate demo, added large file storage for demo files, updated readme and image links, created tests folder for future tests
