+++ sleeppy/sleep.py
@@ -14,6 +14,7 @@ import seaborn as sns
 import struct
 from bitstring import BitArray
 from scipy import signal
+from shutil import copy, rmtree
 
 sns.set()
 pd.options.mode.chained_assignment = None
@@ -72,6 +73,15 @@ class SleepPy:
         stop_buffer="0s",
         start_time="",
         stop_time="",
+        run_config=0,
+        temperature_threshold=25.0,
+        minimum_rest_block=30,
+        allowed_rest_break=60,
+        minimum_rest_threshold=0.0,
+        maximum_rest_threshold=1000.0,
+        minimum_hours=6,
+        clear_intermediate_data=False,
+        aws_object=None,
     ):
         """
         Class initialization.
@@ -83,8 +93,21 @@ class SleepPy:
         :param stop_buffer: number of seconds to ignore from the end of the recording
         :param start_time: time stamp from which to start looking at data (str) format: "%Y-%m-%d %H:%M:%S:%f"
         :param stop_time: time stamp at which to stop looking at data (str) format: "%Y-%m-%d %H:%M:%S:%f"
+        :param run_config: int in range 0-6, dictates which stage of sleeppy to run from (see self.run())
+        :param temperature_threshold: Minimum temperature at which to accept a candidate rest period.
+        :param minimum_rest_block: number of minutes required to consider a rest period valid (int)
+        :param allowed_rest_break: number of minutes allowed to interrupt major rest period (int)
+        :param minimum_rest_threshold: minimum allowed threshold for determining major rest period (float)
+        :param maximum_rest_threshold: maximum allowed threshold for determining major rest period (float)
+        :param minimum_hours: minimum number of hours required to consider a day useable (int)
+        :param clear_intermediate_data: boolean flag to clear all intermediate data
+        :param aws_object: data object to be processed from aws (in place of source file path
         """
-        self.src = input_file  # save input location
+        if aws_object is not None:
+            self.src = aws_object
+        else:
+            self.src = input_file  # save input location
+        self.extension = input_file.split(".")[-1]
         self.dst = results_directory  # save output location
         self.src_name = input_file.split("/")[-1][0:-4]  # save naming convention
         self.sub_dst = (
@@ -101,6 +124,14 @@ class SleepPy:
         self.stop_buffer = stop_buffer
         self.start_time = start_time
         self.stop_time = stop_time
+        self.run_config = run_config
+        self.min_t = temperature_threshold
+        self.minimum_rest_block = minimum_rest_block
+        self.allowed_rest_break = allowed_rest_break
+        self.minimum_rest_threshold = minimum_rest_threshold
+        self.maximum_rest_threshold = maximum_rest_threshold
+        self.minimum_hours = minimum_hours
+        self.clear = clear_intermediate_data
         self.run()  # run the package
 
     def run(self):
@@ -112,29 +143,37 @@ class SleepPy:
             os.mkdir(self.sub_dst)  # set up output directory
         except OSError:
             pass
-        # split the data into 24 hour periods
-        if ".bin" in self.src:
-            self.split_days_geneactiv_bin()
-        elif ".csv" in self.src:
-            self.split_days_geneactiv_csv()
+        if self.run_config <= 0:
+            # split the data into 24 hour periods
+            if ".bin" in self.src:
+                self.split_days_geneactiv_bin()
+            elif ".csv" in self.src:
+                self.split_days_geneactiv_csv()
+        if self.run_config <= 1:
+            # extract the activity index feature
+            self.extract_activity_index()
[…]
