+++ sleeppy/__init__.py
@@ -1,1 +1,2 @@
-from sleeppy.sleep import *
+from sleeppy.sleep import SleepPy, ColeKripke, band_pass_filter, bin2df, activity_index
+from sleeppy.version import __version__
