### 2019-10-21 d920c50
switch matplotlib backend to match jupyter reqs
