### 2019-10-24 081c709
added proper input suggestions for binder demo
