### 2019-10-24 #1663 @elyiorgos
@lheagy @eskofier Thank you for your patience guys, the binder example is up and running! It's accessible from the readme in the demo section, but the following link should work as well:

https://mybinder.org/v2/gh/elyiorgos/sleeppy/master?filepath=example_notebook%2Fexample.ipynb

The environment may take a little while to set up (5-10 minutes maybe), but the version of the demo I have posted has all the data preloaded, so that portion should run quickly. In the results folders the pdfs that are generated don't open in browser for me (no idea why), but if you download them to local you can see them just fine.

Let me know if you have any issues, thanks again!
Yiorg
