### 2019-10-30 12b6c56
Putting sleeppy for python 3 into package form. Updating meta data, documents, demo, tests, etc.
