+++ sleeppy/tests/__init__.py
@@ -1,1 +1,0 @@
-from demo import run_demo
