### 2019-11-04 c22fad1
Updated readme for new implementation, updated folder structures. Updated images based on new implementation. Updated manual visual test to always fail and notify the test runner that a manual test can be run. Switched nans to np.nan. Fixed hline plotting to match data length as opposed to plotting across all 24 hours even if there is no data.
