### 2019-11-05 #1663 @lheagy
:wave: Hi @eskofier, just checking in. Have you had a chance to check out the binder link posted above? This should give you an environment where you can run the examples and play with the code-base.
