### 2019-11-05 #1663 @eskofier
Hi, I will try to have a look the next few days.

Viele Grüße
Bjoern

> Am 05.11.2019 um 18:03 schrieb Lindsey Heagy <notifications@github.com>:
>
> 👋 Hi @eskofier, just checking in. Have you had a chance to check out the binder link posted above? This should give you an environment where you can run the examples and play with the code-base.
>
