+++ requirements.txt
@@ -6,4 +6,4 @@ matplotlib==3.1.1
 scipy==1.3.1
 more_itertools==7.2.0
 numexpr==2.7.0
-tables==3.6.0
+tables==3.6.1
