### 2019-11-07 e1238ba
Updated ipynb for binder example. Updated license link in paper. Switch tests from unittest to pytest. Added conftest file to streamline data loading for tests. Fixed single plot color plotting issue for summary report. parametrized call to major rest period. Updated version to first release version. Updated requirements. Updated readme to reflect changes.
