+++ sleeppy/test/test_utils.py
@@ -1,0 +1,201 @@
+import pytest
+import numpy as np
+import pandas as pd
+from sleeppy.utils import (
+    activity_index,
+    non_overlapping_windows,
+    bin2df,
+    high_pass_filter,
+    load_geneactiv_csv,
+    downsample,
+    make_dummy_data,
+    total_sleep_time,
+    percent_time_asleep,
+    wake_after_sleep_onset,
+    sleep_onset_latency,
+    number_of_wake_bouts,
+)
+
+
+def test_activity_index():
+    input_array = np.array(
+        (
+            [[1, 2, 3], [2, 3, 2], [3, 4, 2], [4, 5, 2], [5, 6, 2], [6, 7, 2]],
+            [[1, 2, 3], [2, 3, 2], [3, 4, 2], [4, 5, 2], [5, 6, 2], [6, 7, 2]],
+            [[1, 2, 3], [2, 3, 2], [3, 4, 2], [4, 5, 2], [5, 6, 2], [6, 7, 2]],
+        )
+    )
+    test_output = activity_index(input_array)
+    expected_output = np.array([[1.33333333], [1.33333333], [1.33333333]])
+    assert np.allclose(test_output, expected_output)
+
+
+def test_non_overlapping_windows():
+    input_df = pd.DataFrame(np.zeros([100, 3]))
+    output = non_overlapping_windows(input_df, 3)
+    test_shape = output.shape
+    expected_shape = (33, 3, 3)
+    assert np.array_equal(test_shape, expected_shape)
+
+
+def test_bin2df(geneactiv_data):
+    input_path = geneactiv_data("bin")
+    expected_path = geneactiv_data("csv")
+    test_df = bin2df(input_path)
+    expected_df = load_geneactiv_csv(expected_path)
+    assert np.allclose(test_df, expected_df)
+
+
+def test_high_pass_filter():
+    sub_array = [
+        [1, 2, 3],
+        [2, 3, 2],
+        [3, 4, 2],
+        [4, 5, 2],
+        [5, 6, 2],
+        [6, 7, 2],
+        [1, 2, 3],
+        [2, 3, 2],
+        [3, 4, 2],
+        [4, 5, 2],
+        [5, 6, 2],
+        [6, 7, 2],
+        [1, 2, 3],
+        [2, 3, 2],
+        [3, 4, 2],
+        [4, 5, 2],
+        [5, 6, 2],
+        [6, 7, 2],
+        [1, 2, 3],
+        [2, 3, 2],
+        [3, 4, 2],
+        [4, 5, 2],
+        [5, 6, 2],
+        [6, 7, 2],
+        [1, 2, 3],
+        [2, 3, 2],
+        [3, 4, 2],
+        [4, 5, 2],
+        [5, 6, 2],
+        [6, 7, 2],
+        [1, 2, 3],
+        [2, 3, 2],
+        [3, 4, 2],
+        [4, 5, 2],
+        [5, 6, 2],
+        [6, 7, 2],
+    ]
+    input_array = np.array((sub_array, sub_array, sub_array))
+    sub_array = [
+        [-0.14760574, -0.14760574, -0.02994791],
+        [0.71247201, 0.71247201, -1.01133389],
+        [1.57024711, 1.57024711, -0.99253225],
+        [2.42570025, 2.42570025, -0.97354249],
+        [3.27881204, 3.27881204, -0.95436413],
+        [4.12956311, 4.12956311, -0.93499668],
+        [-1.02206592, -1.02206592, 0.08456033],
+        [-0.17609447, -0.17609447, -0.89569262],
+        [0.66745804, 0.66745804, -0.87575505],
+        [1.50857217, 1.50857217, -0.8556265],
+        [2.34722848, 2.34722848, -0.83530651],
[…]
