### 2019-11-13 #1663 @eskofier
Dear all,

I tried the demo, but got stuck - I am copying in the output of the Notebook below, and also provide a screenshot what happens when I "Simply copy the url above, and replace /example.ipynb with /demo to load the demo folder.“. How should we proceed?

—————> Notebook output:

Example sleeppy usage: running the demo
This short tutorial serves to confirm that sleeppy installs and runs correctly.
In [1]:

# We first import the demo from sleeppy.
import sleeppy.tests as demo
In [2]:

# next we get the current working directory
import os
print(os.getcwd())
/home/jovyan/example_notebook

In [*]:

# The next step is to run the demo and provide the current working directory as the results directory
demo.run_demo(binder_demo=True)
Please provide a path to a results directory:    /Users/Bjoern/Desktop/Joss

Your previous entry was not appropriate.
It should follow a format similar to /home/jovyan/example_notebook
Please provide a path to a results directory:

Now you can go to the generated results and explore sleeppy's outputs
Simply copy the url above, and replace /example.ipynb with /demo to load the demo folder.

—————> End notebook output

Viele Grüße
Bjoern

> Am 24.10.2019 um 17:16 schrieb Yiorgos Christakis <notifications@github.com>:
>
> @lheagy @eskofier Thank you for your patience guys, the binder example is up and running! It's accessible from the readme in the demo section, but the following link should work as well:
>
> https://mybinder.org/v2/gh/elyiorgos/sleeppy/master?filepath=example_notebook%2Fexample.ipynb
>
> The environment may take a little while to set up (5-10 minutes maybe), but the version of the demo I have posted has all the data preloaded, so that portion should run quickly. In the results folders the pdfs that are generated don't open in browser for me (no idea why), but if you download them to local you can see them just fine.
>
> Let me know if you have any issues, thanks again!
> Yiorg
>
