### 2019-11-13 #1663 @elyiorgos
Hi Bjoern,

Thanks for giving it a shot! The demo failed because the directory you provided wasn't the current working directory. Since the binder environment is one hosted online, if you provide a local directory  it fails and prompts you to use a directory with the proper format and location (usually `/home/jovyan/example_notebook`). That's the reason the current working directory is printed in the previous statement, so it can be copy pasted easily as input.

So just to be 100% clear, when you run the demo and it prompts you with:

 Please provide a path to a results directory:

You should just copy the current working directory printed above, (likely: /home/jovyan/example_notebook)
and continue with that.

Let me know if you have any more issues, thank you!

Best,
Yiorgos
