### 2019-11-13 #1663 @elyiorgos
Hi Bjoern,

Happy to hear it worked!

With respect to the pdf's I think I had mentioned somewhere in an earlier message in the thread that for some reason the pdf's aren't viewable in browser, to see them you have to download them locally unfortunately.

In terms of what happens during the demo/during any given run of the package I would refer you to the readme (or the paper itself) for a more detailed description of what's going on in general, but in short:

- The input data is Geneactiv accelerometer data collected from the wrist location (in either .csv or .bin form, the demo data is 24 hours of sample data in .bin format)
- The output data is a set of quantitative clinical measures related to sleep in csv format, as well as visual reports showcasing those clinical measures both per 24 hour window, and across the entire recording (For the demo there's only a single day's worth).
- The calculation of clinical measures is achieved by the running of several published algorithms (implemented here as methods within the sleeppy class, or as separate classes). These include the calculation of major rest period(the window during which the person is in bed for the night), the calculation of an activity feature (used as an input to sleep/wake classification), the classification of each minute of data to either sleep or wake, and a regime for on-body detection.

With respect to versioning I will chime in and say that the python 3 implementation is essentially finished (barring some validation and documentation of some of the implemented updates), the python 2 implementation will be retired, but frozen, and a more comprehensive and thorough testing suite will accompany the python 3 version (using pytest as a framework).

I hope that helped clear things up, please let me know if you have any more questions. Thank you!

Best,
Yiorgos
