### 2019-11-14 551a818
Update ck scale factor, update resulting demo image
