### 2019-11-19 6c0b1c5
Switched minimum length of major rest period from one to two to prevent an identical start and end time
