### 2019-11-19 acd0e1c
Added a catch for days with no detected major rest period
