### 2019-11-20 ad014ff
Fixed bug where nans affected candidate rest periods at the first 2.5 and last 2.5 minuts of each 24 segment. Made sure major rest period params were getting passed in the sleep class. Added broader plot closing for batch runs. Reset mrp algo params based on pub.
