### 2019-11-25 242993f
Update tests to reflect new implementation change, add plot flag
