### 2019-11-26 #1663 @elyiorgos
Hi all, just checking in to see if there's anything missing or anything I can clear up. Thanks again for taking the time. For those of you celebrating Thanksgiving soon I hope you enjoy the break!

Best,
Yiorgos
