### 2019-11-26 #1663 @eskofier
Dear other reviewers and editor,

I did ask about the version issue - if the others don‘t see this as problematic, Ineould be OK with accepting overall.

Best regards,
Bjoern

-sent from mobile office-

> Am 26.11.2019 um 22:04 schrieb Yiorgos Christakis <notifications@github.com>:
>
> ﻿
> Hi all, just checking in to see if there's anything missing or anything I can clear up. Thanks again for taking the time. For those of you celebrating Thanksgiving soon I hope you enjoy the break!
>
> Best,
> Yiorgos
>
