### 2019-11-27 8898f91
Added sphinx documentation, fixed a file read/write error in tests, added extra documentation to activity index function
