+++ docs/index.rst
@@ -1,0 +1,49 @@
+.. SleepPy documentation master file, created by
+   sphinx-quickstart on Mon Nov 25 10:03:32 2019.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to SleepPy's documentation!
+===================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+SleepPy package
+===============
+
+Submodules
+----------
+
+sleeppy.colekripke module
+-------------------------
+
+.. automodule:: sleeppy.colekripke
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+sleeppy.majorrest module
+------------------------
+
+.. automodule:: sleeppy.majorrest
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+sleeppy.sleep module
+--------------------
+
+.. automodule:: sleeppy.sleep
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+sleeppy.utils module
+--------------------
+
+.. automodule:: sleeppy.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
