### 2019-12-03 #1663 @lheagy
Thanks @eskofier! Have you had a chance to take a look at the check-boxes in the main issue description? If you are satisfied with their current state, could you please tick off the boxes under your name? If you are unable to check the boxes, you might need to accept the invitation at: https://github.com/openjournals/joss-reviews/invitations
