### 2019-12-04 #1663 @eskofier
Hi,

I just read through all the boxes and I would tick them all off if I could - how do I do that (on an iPhone)? I just spent 15 minutes searching and trying.

Best regards,
Bjoern

-sent from mobile office-

> Am 04.12.2019 um 00:56 schrieb Lindsey Heagy <notifications@github.com>:
>
> ﻿
> Thanks @eskofier! Have you had a chance to take a look at the check-boxes in the main issue description? If you are satisfied with their current state, could you please tick off the boxes under your name? If you are unable to check the boxes, you might need to accept the invitation at: https://github.com/openjournals/joss-reviews/invitations
>
