### 2019-12-06 #1663 @lheagy
Thanks @eskofier, with your written confirmation, I can tick them off for you.
