### 2019-12-09 #1663 @yiorg
Thank you everyone!

From what I can tell the article proof looks good. Please let me know about any next steps on our end.

Best,
Yiorgos
