### 2019-12-12 #1663 @elyiorgos
Hi @lheagy,

Based on the above passing I went ahead and made a new release (0.2.21) and archived it at zenodo with DOI: 10.5281/zenodo.3572581

I made sure the title of the archive matches the title of the paper and that the authors of the archive match the authors of the paper.

Please let me know if everything looks good and if there's anything else for me to take care of. Thank you!

Best,
Yiorgos
