### 2019-12-16 #1663 @lheagy
Excellent, thanks @elyiorgos! and apologies for my delay - I was at the AGU conference last week and was a bit overloaded.

This looks good. Many thanks @eskofier, @ikarosilva, @nhejazi for your review 🎉
