### 2019-12-16 #1663 @lheagy
:wave: @openjournals/joss-eics, this submission is ready fo final checks!
