### 2019-12-18 #1663 @arfon
@eskofier, @ikarosilva, @nhejazi - many thanks for your reviews here and to @lheagy for editing this submission ✨

Also, special congratulations to @lheagy for editing her final JOSS paper. Thanks for all of efforts @lheagy making JOSS the success it is!
