### 2019-12-18 #1663 @arfon
> I'm going to keep this issue upon for a while until the PDF is resolving on GitHub pages...

Update - looks like this is working now.
